requirements.txt
-i https://pypi.org/simple
asgiref==3.8.1; python_version >= '3.8'
asttokens==2.4.1
attrs==23.2.0; python_version >= '3.7'
certifi==2024.6.2; python_version >= '3.6'
cffi==1.16.0; platform_python_implementation != 'PyPy'
charset-normalizer==3.3.2; python_full_version >= '3.7.0'
cryptography==42.0.8; python_version >= '3.7'
decorator==5.1.1; python_version >= '3.5'
defusedxml==0.8.0rc2; python_version >= '3.6'
django==5.0.6; python_version >= '3.10'
django-cors-headers==4.3.1; python_version >= '3.8'
django-extensions==3.2.3; python_version >= '3.6'
django-templated-mail==1.1.1
djangorestframework==3.15.1; python_version >= '3.6'
djangorestframework-simplejwt==5.3.1; python_version >= '3.8'
djoser==2.2.3; python_version >= '3.8' and python_version < '4.0'
drf-spectacular==0.27.2; python_version >= '3.7'
drf-spectacular-sidecar==2024.6.1; python_version >= '3.6'
executing==2.0.1; python_version >= '3.5'
idna==3.7; python_version >= '3.5'
inflection==0.5.1; python_version >= '3.5'
iniconfig==2.0.0; python_version >= '3.7'
ipython==8.25.0; python_version >= '3.10'
jedi==0.19.1; python_version >= '3.6'
jsonschema==4.22.0; python_version >= '3.8'
jsonschema-specifications==2023.12.1; python_version >= '3.8'
matplotlib-inline==0.1.7; python_version >= '3.8'
oauthlib==3.2.2; python_version >= '3.6'
packaging==24.1; python_version >= '3.8'
parso==0.8.4; python_version >= '3.6'
pexpect==4.9.0; sys_platform != 'win32' and sys_platform != 'emscripten'
pillow==10.3.0; python_version >= '3.8'
pluggy==1.5.0; python_version >= '3.8'
prompt-toolkit==3.0.47; python_full_version >= '3.7.0'
psycopg2-binary==2.9.3; python_version >= '3.6'
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.22; python_version >= '3.8'
pygments==2.18.0; python_version >= '3.8'
pyjwt==2.8.0; python_version >= '3.7'
pytest==8.2.2; python_version >= '3.8'
pytest-django==4.8.0; python_version >= '3.8'
python3-openid==3.2.0
pytils==0.4.1
pyyaml==6.0.1; python_version >= '3.6'
referencing==0.35.1; python_version >= '3.8'
requests==2.32.3; python_version >= '3.8'
requests-oauthlib==2.0.0; python_version >= '3.4'
rpds-py==0.18.1; python_version >= '3.8'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
social-auth-app-django==5.4.1; python_version >= '3.8'
social-auth-core==4.5.4; python_version >= '3.8'
sqlparse==0.5.0; python_version >= '3.8'
stack-data==0.6.3
traitlets==5.14.3; python_version >= '3.8'
typing-extensions==4.12.2; python_version < '3.12'
uritemplate==4.1.1; python_version >= '3.6'
urllib3==2.2.1; python_version >= '3.8'
wcwidth==0.2.13
